Remove unused feature requirement
We only rely on min_const_generics since there are no compile time
computations based on them. This is stabilized in 1.51 (end of May) and
already stable on nightly. Note that the module itself still IS nightly
since it is not stable to use on the stable compiler but will become
beta soon.