Add more indexing freedom with upper bound
This also moves the ref methods to provide an ExactSize which is more
useful than a pure length. The reverse methods for constructing one now
take the new Capacity struct, which describes an upper bound and is the
anti-struct of Len, which describes a lower bound.

This enables a few cool new patterns for proofs and we also add the
reflexive proofs for equality types.